pyproject.toml
[project]
name = "Pulsarity"
version = "0.1.0-dev"
description = "A demonstrator project for a RotorHazard variant"
authors = [
    { name = "Bryce Gruber", email = "dev@brycegruber.com"}
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11,<4.0"
classifiers = [
    "Framework :: AsyncIO",
    "License :: OSI Approved :: MIT License"
]

dependencies = [
    "starlette (>=0.46.1,<0.47.0)",
    "starsessions (>=2.2.1,<3.0.0)",
    "aiosqlite (>=0.20.0,<0.21.0)",
    "hypercorn (>=0.17.3,<0.18.0)",
    "uvloop (>=0.21.0,<0.22.0) ; sys_platform == \"linux\" or sys_platform == \"darwin\"",
    "winloop (>=0.1.7,<0.2.0) ; sys_platform == \"win32\"",
    "cryptography (>=44.0.1,<45.0.0)",
    "argon2-cffi (>=23.1.0,<24.0.0)",
    "tomlkit (>=0.13.2,<0.14.0)",
    "tortoise-orm (>=0.25.0,<0.26.0)",
    "pydantic (>=2.11.0,<2.12.0)",
    "tool>=0.8.0",
    "mypy>=1.16.1",
]

[project.urls]
repository = "https://github.com/i-am-grub/Pulsarity"

[project.optional-dependencies]
pgsql = [ "asyncpg>=0.30,<0.31" ]

[build-system]
requires = ["uv_build>=0.7.5,<0.8.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
source-include = ["**/py.typed"]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "httpx-ws>=0.7.2",
    "pylint>=3.3.7",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-cov>=6.1.1",
    "pytest-randomly>=3.16.0",
    "tox>=4.26.0",
]
docs = [
    "myst-parser>=4.0.1",
    "sphinx>=8.2.3",
    "sphinx-book-theme>=1.1.4",
    "sphinx-copybutton>=0.5.2",
    "sphinx-substitution-extensions>=2025.6.6",
]

[tool.mypy]
exclude = ["tests"]

[tool.pytest.ini_options]
addopts = "--no-cov-on-fail --showlocals --strict-markers"
asyncio_default_fixture_loop_scope = "function"
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.pylint]
ignore = "tests"

[tool.tox]
requires = ["tox>=4", "tox-uv>=1.26"]
env_list = ["3.13", "3.12", "3.11"]

[tool.tox.env_run_base]
runner = "uv-venv-lock-runner"
description = "run unit tests"
dependency_groups = ["dev"]
commands = [["pytest", {cov="src",  replace = "posargs", default = ["tests"], extend = true }]]
